feat: platform rename (Aion→Praxis), resolve all critical/high defects, add docs chatbot
Platform Rename:
- AION_ env prefix → PRAXIS_, AionBaseError → PraxisBaseError
- Database names, K8s namespace, package name, all docs updated

Critical Defect Fixes (D-1 through D-9):
- Financial precision: all float→Decimal across PnL, tax, execution,
  risk gate, circuit breaker, validation, and backtesting paths
- Migration 009: backtest_results columns → DECIMAL(20,8)
- Stop-loss enforcement: StopLossMonitor in PnL tick processor
- Kill switch: Redis-backed global halt with API endpoints
- Rate limiter: Redis sliding-window replacing stub
- WebSocket reconnection with exponential backoff

Frontend:
- Zoomable/pannable Mermaid diagrams with expand-to-fullscreen
- RAG documentation chatbot (Viking search → Claude streaming)

+++ services/hot_path/src/hitl_gate.py
@@ -165,9 +165,9 @@ class HITLGate:
             return "high_volatility_regime"
 
         # 3. Large trade size relative to allocation
-        max_alloc = float(state.risk_limits.max_allocation_pct)
+        max_alloc = state.risk_limits.max_allocation_pct
         if max_alloc > 0:
-            size_pct = (risk_result.suggested_quantity / max_alloc) * 100
+            size_pct = float(risk_result.suggested_quantity / max_alloc) * 100
             if size_pct > settings.HITL_SIZE_THRESHOLD_PCT:
                 return f"large_trade_{size_pct:.1f}pct"
